@@ -1,18 +1,19 @@
+import { Link } from 'react-router-dom';
 import './appHeader.scss';
 
 const AppHeader = () => {
     return (
         <header className="app__header">
             <h1 className="app__title">
-                <a href="#">
+                <Link to={'/MarvelApp'}>
                     <span>Marvel</span> information portal
-                </a>
+                </Link>
             </h1>
             <nav className="app__menu">
                 <ul>
-                    <li><a href="#">Characters</a></li>
+                    <li><Link to={'/MarvelApp'}>Characters</Link></li>
                     /
-                    <li><a href="#">Comics</a></li>
+                    <li><Link to={'/comics'}>Comics</Link></li>
                 </ul>
             </nav>
         </header>
